chore(pages): update generated types and methods

+++ src/resources/pages/projects/projects.ts
@@ -1,6 +1,7 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 import { APIResource } from '../../../resource';
+import { isRequestOptions } from '../../../core';
 import * as Core from '../../../core';
 import * as DomainsAPI from './domains';
 import {
@@ -47,7 +48,7 @@ export class Projects extends APIResource {
    * ```
    */
   create(params: ProjectCreateParams, options?: Core.RequestOptions): Core.APIPromise<Project> {
-    const { account_id, ...body } = params;
+    const { account_id = this._client.accountId, ...body } = params;
     return (
       this._client.post(`/accounts/${account_id}/pages/projects`, { body, ...options }) as Core.APIPromise<{
         result: Project;
@@ -69,10 +70,18 @@ export class Projects extends APIResource {
    * ```
    */
   list(
-    params: ProjectListParams,
+    params?: ProjectListParams,
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<ProjectsV4PagePaginationArray, Project>;
+  list(options?: Core.RequestOptions): Core.PagePromise<ProjectsV4PagePaginationArray, Project>;
+  list(
+    params: ProjectListParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.PagePromise<ProjectsV4PagePaginationArray, Project> {
-    const { account_id, ...query } = params;
+    if (isRequestOptions(params)) {
+      return this.list({}, params);
+    }
+    const { account_id = this._client.accountId, ...query } = params;
     return this._client.getAPIList(`/accounts/${account_id}/pages/projects`, ProjectsV4PagePaginationArray, {
       query,
       ...options,
@@ -92,10 +101,19 @@ export class Projects extends APIResource {
    */
   delete(
     projectName: string,
-    params: ProjectDeleteParams,
+    params?: ProjectDeleteParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<ProjectDeleteResponse | null>;
+  delete(projectName: string, options?: Core.RequestOptions): Core.APIPromise<ProjectDeleteResponse | null>;
+  delete(
+    projectName: string,
+    params: ProjectDeleteParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<ProjectDeleteResponse | null> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.delete(projectName, {}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return (
       this._client.delete(
         `/accounts/${account_id}/pages/projects/${projectName}`,
@@ -121,7 +139,7 @@ export class Projects extends APIResource {
     params: ProjectEditParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<Project> {
-    const { account_id, ...body } = params;
+    const { account_id = this._client.accountId, ...body } = params;
     return (
       this._client.patch(`/accounts/${account_id}/pages/projects/${projectName}`, {
         body,
@@ -143,10 +161,19 @@ export class Projects extends APIResource {
    */
   get(
     projectName: string,
-    params: ProjectGetParams,
+    params?: ProjectGetParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<Project>;
+  get(projectName: string, options?: Core.RequestOptions): Core.APIPromise<Project>;
+  get(
+    projectName: string,
+    params: ProjectGetParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<Project> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.get(projectName, {}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return (
       this._client.get(`/accounts/${account_id}/pages/projects/${projectName}`, options) as Core.APIPromise<{
         result: Project;
@@ -168,10 +195,22 @@ export class Projects extends APIResource {
    */
   purgeBuildCache(
     projectName: string,
-    params: ProjectPurgeBuildCacheParams,
+    params?: ProjectPurgeBuildCacheParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<ProjectPurgeBuildCacheResponse | null>;
+  purgeBuildCache(
+    projectName: string,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<ProjectPurgeBuildCacheResponse | null>;
+  purgeBuildCache(
+    projectName: string,
+    params: ProjectPurgeBuildCacheParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<ProjectPurgeBuildCacheResponse | null> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.purgeBuildCache(projectName, {}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return (
       this._client.post(
         `/accounts/${account_id}/pages/projects/${projectName}/purge_build_cache`,
@@ -1284,7 +1323,7 @@ export interface ProjectCreateParams {
   /**
    * Path param: Identifier.
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Body param: Name of the project.
@@ -2021,21 +2060,21 @@ export interface ProjectListParams extends V4PagePaginationArrayParams {
   /**
    * Path param: Identifier.
    */
-  account_id: string;
+  account_id?: string;
 }
 
 export interface ProjectDeleteParams {
   /**
    * Identifier.
    */
-  account_id: string;
+  account_id?: string;
 }
 
 export interface ProjectEditParams {
   /**
    * Path param: Identifier.
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Body param: Configs for the project build process.
@@ -2772,14 +2811,14 @@ export interface ProjectGetParams {
   /**
    * Identifier.
    */
-  account_id: string;
+  account_id?: string;
 }
 
 export interface ProjectPurgeBuildCacheParams {
   /**
    * Identifier.
    */
-  account_id: string;
+  account_id?: string;
 }
 
 Projects.ProjectsV4PagePaginationArray = ProjectsV4PagePaginationArray;
